Begin working on styling
