+++ styles/styles.css
@@ -1,0 +1,4 @@
+.nicholas.related-tracks.band-image {
+  height: 50px;
+  width: 40px;
+}
